Apply the fundamental attribution calculations to a portfolio

Calculate the fundamental attribution results for the supplied data. The data is expected to be the output from a fa_load_* function. Portfolio returns are separated into growth, multiple, activity and intra-period components. Portfolio returns and yields implied by the underlying assets are also calculated.

fa_attribution(
  .dt,
  index_col = "index",
  id_col = "id",
  group_cols = c("metric_name", "group_id", "series_name", "series_type"),
  na_rm_start_weights = F
)

Arguments

.dt

A data.table generated by a fa_load_* function. Or, an equivalent data.table/data.frame/tibble containing the returns, yields and starting weights for each asset in the portfolio during each measurement period.

index_col

The character name of the column containing the time index (eg period end date) for each measurement period.

id_col

The character name of the column containing the identifier of each asset in the portfolio.

group_cols

A character vector of columns names identifying each different portfolio and fundamental metric. This is used to group portfolios allowing multiple portfolios and fundamental metrics to be analysed together.

na_rm_start_weights

Remove rows where start_weights are NA. Defaults to FALSE.

Value

A data.table of attribution results.