Prepare attribution results data for plotting or other analysis

This function handles reshaping the attribution results data for use with ggplot2 and cleaning up auto-generated column names so they look nice when included in plot legends etc.

fa_chart_data_prep(
  .dt,
  direction = "long",
  labels = NA,
  extra_id_cols = NA,
  id_cols = c("index", "metric_name", "group_id", "series_name", "series_type"),
  replace_na = c(growth = 0, activity = 0, multiple = 0, intra_period = 0, fund_return =
    0)
)

Arguments

.dt

A data.table of attribution results

direction

Should the prepared data.table be 'long' or 'wide'? Defaults to 'long'.

labels

A named character vector in the form c('intra_period' = 'Intra-Period') to correct column names. Default labels is compatible with default column names from package functions.

extra_id_cols

Additional id cols to be added to id_cols.

id_cols

A character vector of column names that identify each data point by time, portfolio and type (eg strategy, benchmark, relative).

replace_na

Replace NA values in the attribution components with value from a named vector. Defaults to 0. Use NA to retain NAs in the data.

Value

A reshaped and reformatted data.table of attribution results. When reshaping to 'long' the labels for calculated quantities (eg different attribution components) will be in the 'variable' column, and the value in the 'value' column.