Combine asset weights, returns and fundamental-to-price ratios into a data.table

This function combines separate xts objects containing the yields, returns and starting weights of the assets in a portfolio into a single data.table, suitable for the attribution calculations.

fa_load_assets_xts(
  metric_to_price,
  start_weights,
  asset_returns,
  asset_returns_ctb = NULL,
  metric_name = NA,
  series_name = NA,
  group_id = NA
)

Arguments

metric_to_price

An xts containing the fundamental-to-price ratio of the assets in the portfolio. When as asset's weight changes to 0% during a measurement period this ratio is required in the end of the period.

start_weights

An xts containing the starting weights of the assets in the portfolio during the measurement period. For example, if using monthly data the start weights of 31 Jan are the weights of the assets at the beginning of 1 Jan.

asset_returns

An xts containing the returns of the assets during the measurement period.

asset_returns_ctb

An optional xts containing the return contributions of the assets in the portfolio. If not supplied this is calculated as asset_returns * start_weights.

metric_name

A character name of the fundamental being supplied.

series_name

A character name of the strategy from which the starting weights and return contributions are sourced.

group_id

A character name that allows multiple series to be combined, a place holder for when relative attribution is applied.

Value

A data.table of the combined data that can be passed directly to the attribution functions.